+++ layouts/default.vue
@@ -6,21 +6,26 @@ div
   main.mt-4
     .container
       .row
-        .col-12.col-md-3.offset-md-2 
+        .col-12.col-md-3.py-1
           nuxt-link(to="/intensity")
             button.btn.btn-outline-dark.btn-block(
               role="button"
             ) Light Intensity
-        .col-12.col-md-3 
+        .col-12.col-md-3.py-1
           nuxt-link(to="/sensor")
             button.btn.btn-outline-dark.btn-block(
               role="button"
             ) Sensor Size
-        .col-12.col-md-3 
+        .col-12.col-md-3.py-1
           nuxt-link(to="/photometrics")
             button.btn.btn-outline-dark.btn-block(
               role="button"
             ) Light Photometrics
+        .col-12.col-md-3.py-1
+          nuxt-link(to="/file")
+            button.btn.btn-outline-dark.btn-block(
+              role="button"
+            ) File Size
       .pt-4
         <nuxt />
 </template>
@@ -43,5 +48,4 @@ html, body {
   box-sizing: border-box;
   margin: 0;
 }
-
 </style>